# lib/api-zod/src/generated/types/walletResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WalletToken } from './walletToken';

export interface WalletResponse {
  address: string;
  tokens: WalletToken[];
  totalUsdValue: number;
}