# lib/api-zod/src/generated/types/walletToken.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WalletToken {
  mint: string;
  symbol: string;
  name: string;
  logo?: string | null;
  amount: number;
  decimals: number;
  usdValue?: number | null;
  usdPrice?: number | null;
  priceChange24h?: number | null;
}